sql changes

+++ Database_init.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS  crudapp.users (
     login VARCHAR(255) NOT NULL,
     password VARCHAR(255) NOT NULL,
     email VARCHAR(255) NOT NULL,
-    phoneNum BIGINT NOT NULL,
+    phoneNum INT NOT NULL,
     role VARCHAR(50) NOT NULL,
     accCreated DATETIME NOT NULL
 );
